requirements.txt
wget
asyncio
pyrogram
tgcrypto
oauth2client
psycopg2-binary
sqlalchemy==1.3.23
httplib2
pySmartDL
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
youtube_dl
ffmpeg
tenacity
aiohttp
requests
mega.py
python-dotenv
yt_dlp
hachoir
